+++ addons/base/CfgVehicles.hpp
@@ -9,21 +9,6 @@ class CfgVehicles {
                 condition = QUOTE(true);
                 exceptions[] = {"isNotInside", "isNotSitting"};
                 statement = ""; // With no statement the action will only show if it has children
-
-                class GVAR(cell) {
-                    displayName = CSTRING(Cell_Jamming);
-                    condition = QUOTE(true);
-                    statement = ""; // With no statement the action will only show if it has children
-                    priority = 0.5;
-                    insertChildren = QUOTE(_this call FUNC(cell_addChildren));
-                };
-                class GVAR(rf) {
-                    displayName = CSTRING(RF_Jamming);
-                    condition = QUOTE(true);
-                    statement = ""; // With no statement the action will only show if it has children
-                    priority = 0.5;
-                    insertChildren = QUOTE(_this call FUNC(rf_addChildren));
-                };
             };
         };
     };
@@ -42,8 +27,6 @@ class CfgVehicles {
         };
         maximumLoad = 250;
         mass = 150;
-        GVAR(isCellJammer) = QUOTE(true);
-        GVAR(isRFJammer) = QUOTE(true);
     };
 
     // ZEUS Modules
@@ -63,12 +46,5 @@ class CfgVehicles {
                 init = QUOTE(_this call FUNC(initModules));
             };
         };
-    };
-    class GVAR(Cell_ConnectExplosive): GVAR(ModuleBase) {
-        curatorCanAttach = 1;
-        category = "ace_zeus_Utility";
-        displayName = CSTRING(Modules_Cell_ConnectExplosive);
-        function = QFUNC(cell_connectExplosive);
-        icon = "z\ace\addons\explosives\UI\Explosives_Menu_ca.paa";
     };
 };
