+++ addons/base/XEH_PREP.hpp
@@ -1,17 +1,2 @@
-PREP(cell_addChildren);
-PREP(cell_addChildrenChildren);
-PREP(cell_connectExplosive);
-PREP(cell_disableJamming);
-PREP(cell_enableJamming);
-PREP(cell_isJammer);
-PREP(cell_isJamming);
-PREP(handleAceDetonation);
-PREP(handleDetonationAttempted);
 PREP(handleLambsShare);
 PREP(initModules);
-PREP(rf_addChildren);
-PREP(rf_addChildrenChildren);
-PREP(rf_disableJamming);
-PREP(rf_enableJamming);
-PREP(rf_isJammer);
-PREP(rf_isJamming);
