+++ addons/base/XEH_postInit.sqf
@@ -2,7 +2,7 @@
 
 /*
  *	Check for mods
- */
+*/
 
 // ACRE
 GVAR(hasACRE) = false;
@@ -36,12 +36,6 @@ if (isClass(configFile >> "CfgPatches" >> "lambs_main")) then {
 
 
 if (hasInterface) then {
-    // ACE Explosives Jamming
-    [{
-        _this call FUNC(handleAceDetonation)
-    }] call ace_explosives_fnc_addDetonateHandler;
-    [QGVAR(detonationAttempted), {_this call FUNC(handleDetonationAttempted)}] call CBA_fnc_addEventHandler;
-
     // LAMBS AI Information Sharing Jammer
     if (GVAR(hasLAMBS)) then {
         [{
