+++ addons/explosives/$PBOPREFIX$
@@ -1,0 +1,1 @@
+z\jsoc_ew\addons\explosives
