+++ addons/explosives/XEH_preInit.sqf
@@ -1,0 +1,5 @@
+#include "script_component.hpp"
+ADDON = false;
+#include "XEH_PREP.hpp"
+#include "initSettings.sqf"
+ADDON = true;
