+++ addons/explosives/functions/script_component.hpp
@@ -1,0 +1,1 @@
+#include "\z\jsoc_ew\addons\explosives\script_component.hpp"
