+++ addons/explosives/script_component.hpp
@@ -1,0 +1,14 @@
+#define COMPONENT explosives
+#include "\z\jsoc_ew\addons\main\script_mod.hpp"
+
+// #define DEBUG_MODE_FULL
+// #define DISABLE_COMPILE_CACHE
+
+#ifdef DEBUG_ENABLED_EXPLOSIVES
+    #define DEBUG_MODE_FULL
+#endif
+    #ifdef DEBUG_SETTINGS_OTHER
+    #define DEBUG_SETTINGS DEBUG_SETTINGS_EXPLOSIVES
+#endif
+
+#include "\z\jsoc_ew\addons\main\script_macros.hpp"
